fix write session delegation
