+++ agents/agents-core/src/commonMain/kotlin/ai/koog/agents/core/agent/session/AIAgentLLMReadSession.kt
@@ -100,5 +100,4 @@ public expect class AIAgentLLMReadSession internal constructor(
     ): StructuredResponse<T>
 
     override suspend fun requestLLMMultipleChoices(): List<LLMChoice>
-    override fun close()
 }
